Fix "bot not in voice channel" bug
Fix a bug where the bot could be in a voice channel, but not have a
MusicPlayer. This adds a check in cmd_summon to verify that we have a
voice client AND a player for a guild. Probably not the best way to do
it, but the easiest for sure. In the future, we should probably check
for a voice client and a player when seeing if we are connected to a
server.